feat: dashboard update route
