+++ src/components/dashboard/types.ts
@@ -1,0 +1,5 @@
+export type ResumeListItem = {
+  id: string;
+  original_filename: string;
+  created_at: string | null;
+};
